rus-crawler/requirements.txt
elasticsearch
ScrapyElasticSearch
textract
elasticsearch_dsl
